+++ src/delayed-event-queue.ts
@@ -1,5 +1,5 @@
-import { asyncScheduler, Observable, of, Subject } from 'rxjs';
-import { delay, mapTo, mergeMap } from 'rxjs/operators';
+import { Observable, Subject } from 'rxjs';
+import { mapTo, mergeMap } from 'rxjs/operators';
 
 export class DelayedEventQueue {
   private queueArray: Subject<void>[] = [];
@@ -22,19 +22,17 @@ export class DelayedEventQueue {
   }
 
   private signalNext(): void {
-    of(null)
-      .pipe(delay(1, asyncScheduler))
-      .subscribe(() => {
-        const queueLength = this.queueArray.length;
-        // eslint-disable-next-line no-plusplus
-        for (let i = 0; i < queueLength; ++i) {
-          this.queueArray[0].next();
-          this.queueArray.shift();
-        }
-        if (this.queueArray.length > 0) {
-          this.signalNext();
-        }
-      });
+    setTimeout(() => {
+      const queueLength = this.queueArray.length;
+      // eslint-disable-next-line no-plusplus
+      for (let i = 0; i < queueLength; ++i) {
+        this.queueArray[0].next();
+        this.queueArray.shift();
+      }
+      if (this.queueArray.length > 0) {
+        this.signalNext();
+      }
+    }, 0);
   }
 }